Rename reporting module
